refactor: prune and move FunctionalConnectivitySpheres
